クイズ作成コマンドがデータベースに登録する処理を追加
